examples/rest/python-flask-starter/requirements.txt
requests
flask
python_dotenv

